Polish
This commit reviews the document with minor change (typos, sentence
structure) and more important changes that are described below:

I am not sure what the "Work is done on release branches" section is
all about. We do spike on enhancements in private branches and
collaborate on those. Our merge forward process is quite different
though. Issues are fixed in the oldest maintenance branches and merged
forward to master. That way we have a unique commit for a bug fix rather
than having cherry-pick commits for backports.

The Spring Boot CLI only supports Groovy and you can't script an app
using Java or Kotlin.

If that needs to be clarified, the starters are completely optional (as
auto-configuration is). If you add dependencies manually, the
auto-configuration will work just fine.

Concourse is not proprietary, it is open source software released under
ASL v2 (same license as Spring Boot).

I am not sure about what you mean by "release train". A release train
for us is when several Spring Cloud projects are released together with
a release train name (i.e. `Finchely.SR1`). Spring Boot does not use
a release train and the project is released in one go with a "regular"
version scheme (i.e `2.1.1.RELEASE`).

Feel free to dismiss anything that you don't like :-)